pyproject.toml
[tool.poetry]
name = "gamesbot"
version = "0.1.0"
description = ""
authors = ["quiktea <wishymovies@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
asyncpg = "^0.25.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
sentry-sdk = "^1.5.12"
pyboy = "^1.4.7"
pysdl2-dll = "^2.0.20"
jishaku = {git = "https://github.com/Gorialis/jishaku"}
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
